move downloadPlyFilePlaceIn3dViewer to its own file and import it back into thickanalysisbox and startbutton for use
